+++ .gitignore
@@ -1,6 +1,5 @@
 *.swp
 node_modules
-dist
 
 idea/*
 /.idea/
