+++ index.html
@@ -78,9 +78,7 @@
 </div>
 
 
-<script src="src/render-map.js"></script>
-<script src="src/render-bar.js"></script>
-<script src="src/input.js"></script>
+<script src="dist/bundle.js"></script>
 
 <div id="footer">
     <br/><br/>
